requirements.txt
retry==0.9.2
requests==2.21.0
python-slugify==4.0.0
